Add CSharpStreamToFileWriter. Rename CSharpSourceCodeWriter to CSharlLibrarySetToStreamsWriter
